Tiny fix

+++ src/modules/auth/auth.service.ts
@@ -57,6 +57,8 @@ export class AuthService {
   }
 
   async logout(refreshToken: string): Promise<void> {
+    if (!refreshToken) return;
+
     await this.tokenService.removeRefreshToken(refreshToken);
   }
